refactor: add controller package
