+++ src/main/java/racingcar/controller/Utility.java
@@ -1,10 +1,10 @@
-package racingcar.domain;
+package racingcar.controller;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 
 public class Utility {
-    
+
     public static ArrayList<String> nameStringToArray(String nameString) {
         nameString = removeEmptySpaceFromName(nameString);
         return new ArrayList<String>(Arrays.asList(nameString.split(",")));
